pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maven-core"
version = "0.1.0"
description = "A modular, open-source Python framework for building and managing AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Maven Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "llm", "claude", "mcp"]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "uvicorn>=0.30",
    "starlette>=0.38",
    "argon2-cffi>=23.1",
    "pyjwt[crypto]>=2.8",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "types-PyYAML",
]
cloudflare = [
    # Cloudflare SDK dependencies when available
]
aws = [
    "boto3>=1.34",
    "aioboto3>=13.0",
]

[project.urls]
Homepage = "https://github.com/maven-ai/maven-core"
Documentation = "https://maven-ai.github.io/maven-core"
Repository = "https://github.com/maven-ai/maven-core"

[project.entry-points."maven_core.backends.files"]
local = "maven_core.backends.files.local:LocalFileStore"
cloudflare_r2 = "maven_core.backends.files.cloudflare_r2:CloudflareR2FileStore"

[project.entry-points."maven_core.backends.kv"]
memory = "maven_core.backends.kv.memory:MemoryKVStore"
cloudflare_kv = "maven_core.backends.kv.cloudflare_kv:CloudflareKVStore"

[project.entry-points."maven_core.backends.database"]
sqlite = "maven_core.backends.database.sqlite:SQLiteDatabase"
cloudflare_d1 = "maven_core.backends.database.cloudflare_d1:CloudflareD1Database"

[project.entry-points."maven_core.backends.sandbox"]
local = "maven_core.provisioning.local:LocalSandbox"
cloudflare = "maven_core.provisioning.cloudflare:CloudflareSandbox"

[tool.hatch.build.targets.wheel]
packages = ["src/maven_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=maven_core --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]  # Line length handled by formatter
